Project Setup - Section 1
Created the project in Unreal Engine 5 with basic setup. This project will be coded in Visual Studios 2022 using c++. Please note only the c++ source files will be saved to this repository. No content from Unreal Engine will be a part of this repository as the file sizes are way to large.