Animations - Section 3
Created a ShooterAnimInstance class to set variables used to control animation of the character. Created state machine, animation slot and layered blend per bone nodes in the animation blueprint to fine tune the animation behaviors. Added functions in the ShooterCharacter blueprint to allow the character to fire a weapon complete with shooting particles and sound effects. Coded a line trace in the ShooterCharacter blueprint to allow impact and beam particles to spawn when the line trace hits an object. Created a ShooterHUD class to draw crosshairs on the screen. Fixed the line trace so it starts at the gun barrel and ends at the crosshairs to make it possible for the player to correctly aim at targets. Finally added a couple of blendspaces to control straffing so that the character can have natural animation moving sidewards and backwards. About 4 hours of coding!