The Weapon - Section 5
Created Item Class as a general class for pickup items in the game, then created the Weapon Class as a child of the Item Class. Attached a PickupWidget to the weapon item that is visible when the crosshairs trace over it and when inside the Weapon's area sphere. Widget goes invisible when the character steps out of the area sphere or the crosshairs look away from the Weapon. Added variables to Weapon that are bound to the PickupWidget to display the item name, item icon, ammo count, etc. Coded functions that allow for the Character to swap weapons as long as a valid item (Weapon) is under the crosshairs.