Item Interpolation - Section 6
Created two float curves to handle Item interpolation. First curve interpolates an item from the ground to 50 meters or so above the characters head and back to the ground again. The second curve increases the scale of the item when it is above the character's head and reduces it as it drops back to the ground. The purpose is to showcase the weapon that is about to be equipped. Not a ton of coding needed to accomplish this task.