Reloading - Section 7
Downloaded the Amination Startup Packet from the Epic Games Marketplace. Retargeted one of the reloading animations to Belica. Created variables to handle Ammo and created a Ammo Widget so the counts could be updated on the screen as Belica was firing and reloading. Created a comprehensive Reload Montage and added a few notifies in the montage so that the clip of the Weapon would move during the Reload Animation. Also used notifies on the montage to play sound during reloading. Added sounds in the code to play sounds when swapping one Weapon for another.