Ammo Pickups - Section 9
Created the Ammo class which is a child of the Item class to handle behavior of Ammo as the Character Equips it. Character can equip Ammo either by overlapping the AreaSphere and pressing 'E' or by simply running over the AmmoCollisionSphere. Created a widget for the Ammo that shows the Ammo Icon and Ammo Count whenever the Character overlaps the AreaSphere. Created 6 different interpolation locations that Ammo will interp to at random when Character picks up ammo. Also added functions to limit the number of sound played when multiple Ammo is picked up at once.