Outline & Glow Effects - Section 10
Created a post processing material from scratch to be used on Items in the game. Used this custom post processing material for weapons to glow at a certain color depending on their Item Rarity. Created an onscreen Inventory Widget to be able to store up to 5 different Weapons that are picked up by the character. Player can use the numbers on the widget to equip any weapon in the Inventory. Created a data table to hold features about Items including glow color, custom stencil depth, background color and post process highlight. Variables were set on the Item class to link them to variables in the Data Table. Each Item's pickup widget color is also linked to the data table to match the glowing colors on each item found in the game.