Multiple Weapon Types - Section 11
Created a Weapon Data Table to handle all the properties of the three weapons including the Submachine Gun, Assault Rifle and the Pistol. Created FABRIK (Forward and Backward Reaching Inverse Kinematics), glow materials,  reload animations, animation blueprints, and unique aiming crosshairs for each of the three weapons all of which are read from the data table using c++. Finally setup a boolean for automatic gun fire which is true for SMG and AR, but false for the pistol.