The Enemy Class - Section 13
Created an Enemy class (as a child of the Character class) and downloaded Grux from the Unreal Engine Marketplace for the Skeletal Mesh. Also created a BulletHit interface to be implemented on Enemy and Item classes. For example, Enemies take damage when hit while explosives, well...explode. (Eventually explosives will cause damage to Character classes, but haven't added that yet. Created an EnemyHit Montage for the Enemy to react when being hit by bullets. Added a HitNumber widget to the enemy that shows the actual damage when hit by bullets. Added another Widget to the Enemy class that monitors the Enemy's health as a Progress Bar and reduces when damage is taken. Finally, I added animation to the HitNumber widget so the numbers float away from the enemy and disappear. Programmed head shots to be slighter high damage than body shots.