AI and Behavior Trees - Section 15
Created a Blackboard and Behavior Trees using an AI Controller to control the behavior of the Enemy including patrolling, chasing and attacking the Character when aggravated, being stunned by the Character and dying. Also setup systems that do damage to both Character and Enemy when one is attacking the other. Finally I added "new" enemy types which are really just the Grux character when different skins, sizes and statistics. Far more blueprint coding than c++ which is common when programming animations in Unreal Engine. Next commit will focus on the final Enemy Boss (Khaimera).