Khaimera - Section 16
Using the EnemyBP as a rough starting point, I created a Khaimera blueprint as a boss type Enemy and used that for the different skins. Added attack, death and hit react montages for the Khaimera Enemy. Coded the explosives to cause damage to any Actor (Enemy or Character) within the overlap sphere. Also added a health pickup Item to give the Character a boost in health points (not to exceed maximum health) whenever the character enters the collision sphere. Only left to do now is to build the level. Little to no coding to should left for this project!